requirements.txt
# Core LLM & LangChain integrations
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.6

# OpenAI-compatible client (needed for langchain_openai or fallback if needed)
openai>=1.30.1

# Flashcards & translation
python-dotenv>=1.0.1
deep-translator>=1.11.4

# Web UI
gradio>=4.29.0

# Progress tracking / logging
tqdm>=4.66.2
